+++ dav_utils/__init__.py
@@ -1,0 +1,16 @@
+# -*- coding: utf-8 -*-
+# TODO: подключить doctest
+# TODO: readme
+# TODO: расширить тест на запись в файл - для 1го и нескольких значений
+# TODO: попробовать CI/CD с запуском тестов для коммитов
+# TODO: собрать в пакет для возможной установки через pip
+# TODO: отключить флаг "Шаблон" на github
+
+"""Set of tools that I often have to reproduce in pure Python scripts.
+
+TODO: add modules descriptions and examples.
+"""
+__version__ = '0.2.0'
+
+
+__author__ = 'Aleksey Devyatkin <devyatkin.av@ya.ru>'
